Put up all the routes and controllers in proper directory structure
