+++ index.js
@@ -2,16 +2,14 @@ const express = require('express')
 const app = express()
 var cmd = require('node-cmd')
 const port = 8000
-app.get('/', (req, res) => {
-    var pyProcess = cmd.get('',
-    function(data, err, stderr){
-    if (!err)
-        res.send(data);
-    else 
-        console.log("python script cmd error: " + err)
-    }
-);
-});
 
-app.listen(port, () => console.log(`Example app listening on port 
-${port}!`))
+app.use('/',require('./WebServer/routes'));
+
+app.listen(port,function(err){
+    if(err)
+    {
+       // console.log("Error: ",err);
+        console.log(`Error in running the server: ${err}`);              
+    }
+    console.log(`server is running on port: ${port}`);
+})
